+++ docs/source/changelog.rst
@@ -3672,3 +3672,6 @@ Current C++/SQLite client, Python/SQLAlchemy server
 
 - **New task:** :ref:`Internet Severity and Activities Addiction Questionnaire (ISAAQ)
   <isaaq>`. (Database revision 0073.)
+
+- **New task:** :ref:`Internet Severity and Activities Addiction Questionnaire, Easting Disorders Appendix (ISAAQ-ED)
+  <isaaqed>`. (Database revision 0074.)